% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{getCoordinate}
\alias{getCoordinate}
\title{Funzione che restituisce le cordinate di un indirizzo utilizzando il servizio mappe di regione toscana nel caso di non riuscita utilizza OpenRuoteService}
\usage{
getCoordinate(indirizzo, civico, comune, provincia, siglaprovincia, url, token)
}
\value{
vettore con longitudie e latitudine
NA se l'indirizzo non viene decodificato
}
\description{
Funzione che restituisce le cordinate di un indirizzo utilizzando il servizio mappe di regione toscana nel caso di non riuscita utilizza OpenRuoteService
}
\examples{
indirizzo <- 'via di novoli'
civico <- '26'
comune <- 'firenze'
provincia <- 'Firenze'
siglaprovincia <- 'FI'
URLRT <- 'http://mappe.regione.toscana.it/legacy-musume/services/Musume?method=richiesta&algoritmo=1&'
token <- '5b3ce359785111000we624823fe6724fd45c3681e64432745a3bec'
coord <- getCoordinate(indirizzo,civico,comune,provincia,siglaprovincia,url,token)

}
