% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{getCoordORS}
\alias{getCoordORS}
\title{Funzione che restituisce le cordinate di un indirizzo utilizzando il servizio openruoteservice}
\usage{
getCoordORS(indirizzo, civico, comune, provincia, token)
}
\value{
vettore con longitudie e latitudine
NOCOORDORS se l'indirizzo non viene decodificato
Messagge in caso di errore o warning
}
\description{
Funzione che restituisce le cordinate di un indirizzo utilizzando il servizio openruoteservice
}
\examples{
indirizzo <- 'via di novoli'
civico <- '26'
comune <- 'firenze'
provincia <- 'Firenze'
siglaprovincia <- 'FI'
token <- '5b3ce359785111000we624823fe6724fd45c3681e64432745a3bec'
coordORS <- getCoordORS(indirizzo,civico,comune,provincia,token)

}
