% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{getCoordRT}
\alias{getCoordRT}
\title{Funzione che restituisce le cordinate di un indirizzo utilizzando il servizio mappe della regione toscana}
\usage{
getCoordRT(indirizzo, civico, comune, siglaprovincia, url)
}
\value{
vettore con longitudie e latitudine e una stringa che contiene:
\itemize{
\item la url utilizzata per ottenere le coordinati se vengono trovate
\item NOOKResponse se la chiamata della url ha dato errore
\item NOCOORD se la riposta ha dato esito positivo ma l'indirizzo non è interpretabile
}
}
\description{
Funzione che restituisce le cordinate di un indirizzo utilizzando il servizio mappe della regione toscana
}
\examples{
indirizzo <- 'via di novoli'
civico <- '26'
comune <- 'firenze'
provincia <- 'Firenze'
siglaprovincia <- 'FI'
URLRT <- 'http://mappe.regione.toscana.it/legacy-musume/services/Musume?method=richiesta&algoritmo=1&'
coordRT <- getCoordRT(indirizzo,civico,comune,siglaprovincia,URLRT)

}
